Remove useless ids_ps_advice.
This variable contained the list of all advices including premium and except addons. It was initialized in the hook event, then passed along to AdminGamificationController, who then used it as whitelist for filtering a list of advices of its own. This controller-generated list, before being filtered, will have either _only_ premium advices or all advices _except_ premium and addons. Following this, we can conclude that controller-generated list will always be a fully contained subset of the whitelist, therefore making the whitelist redundant and useless.